pyproject.toml
[build-system]
requires = ["hatchling>=1.18"]
build-backend = "hatchling.build"

[project]
name = "music-normalizer"
version = "0.1.0"
description = "Review-first CLI for normalizing messy TITLE/ARTIST tags on niche psytrance / darkpsy libraries."
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Tal Arbetov" }]
license = { text = "MIT" }
keywords = ["music", "metadata", "mutagen", "ollama", "psytrance"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Topic :: Multimedia :: Sound/Audio",
]

dependencies = [
  "typer>=0.12",
  "rich>=13.7",
  "pydantic>=2.6",
  "mutagen>=1.47",
  "httpx>=0.27",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.5",
  "mypy>=1.10",
]

[project.scripts]
music-normalizer = "music_normalizer.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/music_normalizer"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
